requirements.txt
numpy
gym==0.10.11
imageio
PILLOW
pyglet
pyvirtualdisplay
tf-agents-nightly
tf-nightly
tfp-nightly

matplotlib

pyyaml
pygame